pyproject.toml
[project]
name = "ultrawork"
version = "0.1.0"
description = "Automated task management and approval system with Slack MCP integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "slack_sdk>=3.0",
    "python-dotenv>=1.0",
    "anthropic>=0.76.0",
    "google-genai>=1.60.0",
    "tweepy>=4.16.0",
    "requests>=2.0",
    "google-generativeai>=0.8.6",
    "textual>=7.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
]

[project.scripts]
ultrawork = "ultrawork.main:app"
ultrawork-setup = "ultrawork.installer.app:run_setup"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ultrawork"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "SIM117", "SIM102"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
